move the QuestRegistry and processors into sub-packages
