+++ src/main/java/org/betonquest/betonquest/quest/registry/processor/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides the functionality of the BetonQuest types used in the API.
+ */
+package org.betonquest.betonquest.quest.registry.processor;
